fix: moved files to a folder that is not git ignored
