+++ public/app/deploys/confirm-deploy-view.js
